Revert "0908"
This reverts commit 527ff76bfe07b5752e1a820e46b72a59d7cc8361.

+++ Bossinfo.HealthPlatform.Models/Utility/MIData.cs
@@ -125,6 +125,6 @@ namespace Bossinfo.HealthPlatform.Models.Utility
         /// <summary>
         /// 來源內部序號
         /// </summary>
-        public decimal Source_SN { get; set; }
+        public int Source_SN { get; set; }
     }
 }
